requirements.txt
flask>=3.0.0
curl-cffi>=0.6.0
python-dotenv>=1.0.0
waitress>=2.1.2
